+++ src/mitll/langtest/shared/instrumentation/SlimEvent.java
@@ -6,11 +6,11 @@ import com.google.gwt.user.client.rpc.IsSerializable;
  * Created by go22670 on 4/7/16.
  */
 public class SlimEvent implements IsSerializable, Comparable<SlimEvent> {
-  private long userID;
+  private int userID;
   private long timestamp;
 
   public SlimEvent() {}
-  public SlimEvent(long userID, long timestamp) {
+  public SlimEvent(int userID, long timestamp) {
     this.userID = userID;
     this.timestamp = timestamp;
   }
@@ -20,7 +20,7 @@ public class SlimEvent implements IsSerializable, Comparable<SlimEvent> {
     return timestamp < o.timestamp ? -1 : timestamp > o.timestamp ? +1 : 0;
   }
 
-  public long getUserID() {
+  public int getUserID() {
     return userID;
   }
